==> 0_alternate_dataset_review.ipynb <==
Source: https://www.kaggle.com/datasets/keeschristiaanse/lyrics-with-sentiment

import pandas as pd
import numpy as np

from google.colab import drive
drive.mount('/content/gdrive')

# note: place shortcut to shared project folder in google drive root directory
%cd /content/gdrive/MyDrive/ml2_project

df = pd.read_csv('data/new_set_lyrics_with_sentiment.csv')

df

df.shape

df['year'].describe()

df['genre'].value_counts()

df['artist'].value_counts()

df = df[df['genre'].isin(['Rock','Country'])]
df

import random

songs = random.sample(range(0, len(df)), 10)

print(df['song'].iloc[songs[0]])
print(df['artist'].iloc[songs[0]])
print(df['genre'].iloc[songs[0]])
print(df['lyrics'].iloc[songs[0]])

print(df['song'].iloc[songs[1]])
print(df['artist'].iloc[songs[1]])
print(df['genre'].iloc[songs[1]])
print(df['lyrics'].iloc[songs[1]])

print(df['song'].iloc[songs[2]])
print(df['artist'].iloc[songs[2]])
print(df['genre'].iloc[songs[2]])
print(df['lyrics'].iloc[songs[2]])

print(df['song'].iloc[songs[3]])
print(df['artist'].iloc[songs[3]])
print(df['genre'].iloc[songs[3]])
print(df['lyrics'].iloc[songs[3]])

print(df['song'].iloc[songs[4]])
print(df['artist'].iloc[songs[4]])
print(df['genre'].iloc[songs[4]])
print(df['lyrics'].iloc[songs[4]])

print(df['song'].iloc[songs[5]])
print(df['artist'].iloc[songs[5]])
print(df['lyrics'].iloc[songs[5]])

print(df['song'].iloc[songs[6]])
print(df['artist'].iloc[songs[6]])
print(df['lyrics'].iloc[songs[6]])

print(df['song'].iloc[songs[7]])
print(df['artist'].iloc[songs[7]])
print(df['lyrics'].iloc[songs[7]])

print(df['song'].iloc[songs[8]])
print(df['artist'].iloc[songs[8]])
print(df['lyrics'].iloc[songs[8]])

print(df['song'].iloc[songs[9]])
print(df['artist'].iloc[songs[9]])
print(df['lyrics'].iloc[songs[9]])
