==> 3_final_train.ipynb <==
import pandas as pd
import numpy as np

from google.colab import drive
drive.mount('/content/gdrive')

# check system specs

gpu_info = !nvidia-smi
gpu_info = '\n'.join(gpu_info)
if gpu_info.find('failed') >= 0:
  print('Not connected to a GPU')
else:
  print('Connected to a GPU')

from psutil import virtual_memory
ram_gb = virtual_memory().total / 1e9

if ram_gb < 20:
  print('Not using a high-RAM runtime: {:.1f} gigabytes of available RAM'.format(ram_gb))
else:
  print('Using a high-RAM runtime: {:.1f} gigabytes of available RAM'.format(ram_gb))

# note: place shortcut to shared project folder in google drive root directory
%cd /content/gdrive/MyDrive/ml2_project

df = pd.read_csv('data/clean_dataset.csv')
df = df[df['tag'] != 'rap']
lyrics = df['lyrics']

lyrics.shape

# remove duplicates from list and preserve order
def dedupe(seq):
    seen = set()
    seen_add = seen.add
    return [x for x in seq if not (x in seen or seen_add(x))]

# process dataset

lyrics_deduped = []

for song in lyrics:
    song = dedupe(str(song).split('\n'))
    song = '\n'.join(song)
    lyrics_deduped.append(song)

### Final Train
Train on entire dataset

!pip install sentence_transformers

# prep training data

from sentence_transformers import InputExample

train_samples = []
for song in lyrics_deduped:
    # split songs into two halves for positive pair training
    song = str(song)
    half_1 = song[:len(song)//2]
    half_2 = song[len(song)//2:]
    train_samples.append(InputExample(
        texts=[half_1, half_2]
    ))

from sentence_transformers import datasets

batch_size = 32

# removes duplicate pairings
loader = datasets.NoDuplicatesDataLoader(
    train_samples, batch_size=batch_size)

from sentence_transformers import models, SentenceTransformer
import torch

# model = SentenceTransformer('all-MiniLM-L12-v2')
# model.save('./all-MiniLM-L12-v2')
model = SentenceTransformer('./models/all-MiniLM-L12-v2')
model.max_seq_length = 256
model = model.to(torch.device('cuda')) # use GPU

model

# specify loss

from sentence_transformers import losses

# mnr loss
loss = losses.MultipleNegativesRankingLoss(model)

# model name for export/save
model_export_name = './models/finetune_mnr_256'

# train and export model

epochs = 1
warmup_steps = int(len(loader) * epochs * 0.1)

model.fit(
    train_objectives=[(loader, loss)],
    epochs=epochs,
    warmup_steps=warmup_steps,
    output_path=model_export_name,
    show_progress_bar=True
)

model
