==> 5_semantic_search.ipynb <==
!pip install sentence_transformers

import pandas as pd
import numpy as np
from google.colab import drive #file handling
from sentence_transformers import SentenceTransformer

# check system specs

gpu_info = !nvidia-smi
gpu_info = '\n'.join(gpu_info)
if gpu_info.find('failed') >= 0:
  print('Not connected to a GPU')
else:
  print('Connected to a GPU')

from psutil import virtual_memory
ram_gb = virtual_memory().total / 1e9

if ram_gb < 20:
  print('Not using a high-RAM runtime: {:.1f} gigabytes of available RAM'.format(ram_gb))
else:
  print('Using a high-RAM runtime: {:.1f} gigabytes of available RAM'.format(ram_gb))

Note: shouldn't need high performance runtime

drive.mount('/content/gdrive')

# note: place shortcut to shared project folder in google drive root directory
%cd /content/gdrive/MyDrive/ml2_project

base_model = SentenceTransformer('./models/all-MiniLM-L12-v2')
# base_model = base_model.to(torch.device('cuda')) # use GPU

ft_final_model = SentenceTransformer('./models/finetune_mnr_final')
# ft_final_model = ft_final_model.to(torch.device('cuda')) # use GPU

sds = pd.read_csv('data/small_dataset_w_embeddings.csv')
sds = sds[sds['tag'] != 'rap'].reset_index()

df = pd.read_csv('data/clean_dataset_w_embeddings.csv').drop(labels = ['Unnamed: 0', 'level_0', 'Unnamed: 0.1', 'index'],
                                                             axis = 1)
df = df[df['tag'] != 'rap'].reset_index()

def embedding_str_to_np(embed_as_str):
  #remove brackets and collapse whitespace
  regularized_str = embed_as_str.replace('\n ', ' ').replace("  ", " ")[1:-1]
  regularized_str = regularized_str.strip()
  np_array = np.array(regularized_str.split(" ")).astype(float)
  return(np_array)

sds['embeddings_base'] = sds['embeddings_base'].map(embedding_str_to_np)
sds['embeddings_ft_final'] = sds['embeddings_ft_final'].map(embedding_str_to_np)

df['embeddings_base'] = df['embeddings_base'].map(embedding_str_to_np)
df['embeddings_ft_final'] = df['embeddings_ft_final'].map(embedding_str_to_np)

Because our model is trained to match song halves, and we generally use shorter queries than half a song, our model wants to only return short songs. 

def n_tokens(s):
  return len(ft_final_model.tokenizer(s)['input_ids'])
  
df["num_tokens"] = df['lyrics'].map(n_tokens)

df = df[df['num_tokens'] >= 128]

def cosine_similarity(v1, v2):
    d = np.dot(v1, v2)
    cos_theta = d / (np.linalg.norm(v1) * np.linalg.norm(v2))
    return(cos_theta)

def relevance_scores(query_embed,df,embeddings):
    scores = [cosine_similarity(query_embed, v2) for v2 in df[embeddings]]
    scores = pd.Series(scores)
    return(scores)

def semantic_search(query_sentence, df = df, model = ft_final_model, embeddings = 'embeddings_ft_final', return_top = False):
    df = df.copy()
    query_embed = model.encode(query_sentence)
    scores = relevance_scores(query_embed,df,embeddings)
    df['scores'] = scores
    sorted_df = df.sort_values(by = 'scores', ascending = False)

    #def no_newlines(s):
    #  return s.replace("\n",", ")
    #sorted_df['lyrics'] = sorted_df['lyrics'].map(no_newlines)

    if return_top == False:
        return sorted_df[['title','artist','lyrics','scores']]
    else:
        return sorted_df.iloc[0]['lyrics']

pd.set_option('display.width', 100)

q = "I’m happy sometimes but also sad. You wouldn’t understand"

semantic_search(q, sds, base_model, 'embeddings_base')

semantic_search(q, df, ft_final_model, 'embeddings_ft_final')

print(semantic_search(q, df, ft_final_model, 'embeddings_ft_final', return_top = True))
